+++ docs/guides/release-notes.md
@@ -12,9 +12,19 @@ Track all ZPA Terraform provider's releases. New resources, features, and bug fi
 
 ---
 
-``Last updated: v4.0.10``
+``Last updated: v4.0.11``
 
 ---
+
+## 4.0.11 (April, 14 2025)
+
+### Notes
+
+- Release date: **(April, 14 2025)**
+- Supported Terraform version: **v1.x**
+
+### Bug Fixes
+[PR #548](https://github.com/zscaler/zscaler-sdk-go/pull/548) - Set pointer in the `credential` block attribute in the ZPA `policysetcontrollerv2` resource.
 
 ## 4.0.10 (April, 11 2025)